+++ core/src/main/java/org/vertexium/GraphBaseWithSearchIndex.java
@@ -18,12 +18,14 @@ public abstract class GraphBaseWithSearchIndex extends GraphBase implements Grap
     private boolean foundIdGeneratorClassnameInMetadata;
 
     protected GraphBaseWithSearchIndex(GraphConfiguration configuration) {
+        super(configuration.isStrictTyping());
         this.configuration = configuration;
         this.searchIndex = configuration.createSearchIndex(this);
         this.idGenerator = configuration.createIdGenerator(this);
     }
 
     protected GraphBaseWithSearchIndex(GraphConfiguration configuration, IdGenerator idGenerator, SearchIndex searchIndex) {
+        super(configuration.isStrictTyping());
         this.configuration = configuration;
         this.searchIndex = searchIndex;
         this.idGenerator = idGenerator;